+++ public/video-ui/src/components/VideoPoster/VideoPoster.js
@@ -2,7 +2,7 @@ import React from 'react';
 import VideoPosterEdit from '../VideoEdit/formComponents/VideoPoster';
 import validate from '../../constants/posterEditValidation';
 import {ManagedForm, ManagedField} from '../ManagedForm';
-import FormImageSelector from '../FormFields/FormImageSelector';
+import ImageSelector from '../FormFields/ImageSelector';
 
 export default class VideoPoster extends React.Component {
 
@@ -17,7 +17,7 @@ export default class VideoPoster extends React.Component {
           fieldLocation="posterImage"
           name=""
         >
-          <FormImageSelector editMode={true}/>
+          <ImageSelector editMode={true}/>
         </ManagedField>
       </ManagedForm>
     );
